fix(ribbon): consolidate types to resolve export ambiguity
Issue: #330

**Problem:** RibbonTab interface was defined in both mp-ribbon.element.ts
and ribbon.component.ts, causing TypeScript export conflict:
'Module has already exported a member named RibbonTab'.

**Solution:**
- Create shared types file: src/lib/types/ribbon.types.ts
- Export RibbonTab + RibbonTabChangeEvent from types module
- Import types in both WC and Angular component
- Update index.ts to explicitly export types (not via re-export)
- Update demo component to use proper type imports

**Changes:**
- New file: libs/mintplayer-ng-bootstrap/ribbon/src/lib/types/ribbon.types.ts
- Modified: mp-ribbon.element.ts (import types)
- Modified: ribbon.component.ts (import types)
- Modified: index.ts (explicit exports to avoid conflicts)
- Modified: demo ribbon.component.ts (use proper imports)

This consolidation also improves maintainability: types are now in a
single location and can be extended without duplication.
