+++ libs/mintplayer-ng-bootstrap/ribbon/src/index.ts
@@ -1,9 +1,12 @@
+// Types
+export type { RibbonTab, RibbonTabChangeEvent } from './lib/types/ribbon.types';
+
 // Angular components (wrappers)
-export * from './lib/components/ribbon.component';
-export * from './lib/components/ribbon-group.component';
-export * from './lib/components/ribbon-button.component';
+export { BsRibbonComponent } from './lib/components/ribbon.component';
+export { BsRibbonGroupComponent } from './lib/components/ribbon-group.component';
+export { BsRibbonButtonComponent } from './lib/components/ribbon-button.component';
 
 // Lit WC (core logic)
-export * from './lib/web-components/mp-ribbon.element';
-export * from './lib/web-components/mp-ribbon-group.element';
+export { MpRibbon } from './lib/web-components/mp-ribbon.element';
+export { MpRibbonGroup } from './lib/web-components/mp-ribbon-group.element';
 export * from './lib/web-components/items';
