+++ libs/mintplayer-ng-bootstrap/ribbon/src/lib/types/ribbon.types.ts
@@ -1,0 +1,10 @@
+export interface RibbonTab {
+  id: string;
+  label: string;
+  content?: string;
+}
+
+export interface RibbonTabChangeEvent {
+  previousTabId: string;
+  activeTabId: string;
+}
